% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing.R
\name{fishing}
\alias{fishing}
\title{Reading A Dataframe From The datalake API}
\usage{
fishing(datafileid, millesime = NULL)
}
\arguments{
\item{datafileid}{the unique identifier of the individual dataframe to read into R.}

\item{millesime}{the millesime of the datafile. Default is the last one}
}
\value{
a dataframe
}
\description{
This function read directly into R a specific datafile by the mean of its unique ID and millesime.
}
\examples{
\dontrun{
fishing(datafileid = "6d031ea5-d1e0-4bd6-82de-1bc3e14acdf2", millesime = "2019-06")}
}
